test(generation): move decoder tests in their own file
